+++ benches/min_distances_unweighted/graph_1.rs
@@ -2,7 +2,7 @@ use {
     divan::Bencher,
     graaf::algo::{
         bfs,
-        dijkstra::unweighted as dijkstra,
+        dijkstra,
     },
 };
 
@@ -10,22 +10,18 @@ fn main() {
     divan::main();
 }
 
-fn graph() -> Vec<Vec<usize>> {
-    vec![
-        vec![1, 3],
-        vec![0, 2],
-        vec![1],
-        vec![0, 4, 7],
-        vec![3, 5, 6, 7],
-        vec![4, 6],
-        vec![4, 5, 7],
-        vec![3, 4, 6],
-    ]
-}
-
 #[divan::bench(min_time = 1)]
 fn dijkstra(bencher: Bencher) {
-    let graph = graph();
+    let graph = vec![
+        vec![(1, 1), (3, 1)],
+        vec![(0, 1), (2, 1)],
+        vec![(1, 1)],
+        vec![(0, 1), (4, 1), (7, 1)],
+        vec![(3, 1), (5, 1), (6, 1), (7, 1)],
+        vec![(4, 1), (6, 1)],
+        vec![(4, 1), (5, 1), (7, 1)],
+        vec![(3, 1), (4, 1), (6, 1)],
+    ];
 
     bencher.bench_local(|| {
         dijkstra::min_distances_single_source(&graph, 0);
@@ -34,7 +30,16 @@ fn dijkstra(bencher: Bencher) {
 
 #[divan::bench(min_time = 1)]
 fn bfs(bencher: Bencher) {
-    let graph = graph();
+    let graph = vec![
+        vec![1, 3],
+        vec![0, 2],
+        vec![1],
+        vec![0, 4, 7],
+        vec![3, 5, 6, 7],
+        vec![4, 6],
+        vec![4, 5, 7],
+        vec![3, 4, 6],
+    ];
 
     bencher.bench_local(|| {
         bfs::min_distances_single_source(&graph, 0);