+++ src/algo/dijkstra/mod.rs
@@ -1,5 +1,0 @@
-/// Dijkstra's algorithm with binary-heap for unweighted graphs
-pub mod unweighted;
-
-/// Dijkstra's algorithm with binary-heap for weighted graphs
-pub mod weighted;